Midway point for refactoring
Next step: tackle tempalte settings, and add a keyValueMaybeFromFile
type to flags.
