+++ cmd/glaze/cmds/docs.go
@@ -41,7 +41,7 @@ var DocsCmd = &cobra.Command{
 
 func init() {
 	DocsCmd.Flags().SortFlags = false
-	cli.AddOutputFlags(DocsCmd, cli.NewOutputFlagsDefaults())
+	cli.AddOutputFlags(DocsCmd)
 	cli.AddTemplateFlags(DocsCmd, cli.NewTemplateFlagsDefaults())
 	cli.AddFieldsFilterFlags(DocsCmd, &cli.FieldsFilterFlagsDefaults{
 		Fields:      "path,Title,SectionType,Slug,Commands,Flags,Topics,IsTopLevel,ShowPerDefault",
