+++ pkg/cli/flags/template.yaml
@@ -1,0 +1,15 @@
+- name: template
+  type: string
+  help: Go Template to use for single string
+  default: ""
+
+- name: template-field
+  type: stringList
+  help: For table output, fieldName:template to create new fields, or @fileName to read field templates from a yaml dictionary
+  default: []
+
+- name: use-row-templates
+  type: bool
+  help: Use row templates instead of column templates
+  default: false
+
