+++ pkg/cli/settings.go
@@ -1,118 +1,12 @@
 package cli
 
 import (
-	"github.com/Masterminds/sprig"
-	"github.com/go-go-golems/glazed/pkg/cmds"
 	"github.com/go-go-golems/glazed/pkg/formatters"
-	"github.com/go-go-golems/glazed/pkg/helpers"
 	"github.com/go-go-golems/glazed/pkg/middlewares"
 	"github.com/go-go-golems/glazed/pkg/types"
-	"github.com/pkg/errors"
 	"gopkg.in/yaml.v3"
 	"os"
-	"text/template"
-	"unicode/utf8"
 )
-
-type TemplateFormatterSettings struct {
-	TemplateFuncMaps []template.FuncMap
-	AdditionalData   map[string]interface{} `glazed.parameter:"template-data"`
-}
-
-type OutputFormatterSettings struct {
-	Output                    string `glazed.parameter:"output"`
-	TableFormat               string `glazed.parameter:"table-format"`
-	OutputAsObjects           bool   `glazed.parameter:"output-as-objects"`
-	FlattenObjects            bool   `glazed.parameter:"flatten-objects"`
-	WithHeaders               bool   `glazed.parameter:"with-headers"`
-	CsvSeparator              string `glazed.parameter:"csv-separator"`
-	Template                  string `glazed.parameter:"template"`
-	TemplateFormatterSettings *TemplateFormatterSettings
-}
-
-func NewOutputFormatterSettings(parameters map[string]interface{}) (*OutputFormatterSettings, error) {
-	s := &OutputFormatterSettings{}
-	err := cmds.InitializeStructFromParameters(s, parameters)
-	if err != nil {
-		return nil, errors.Wrap(err, "failed to initialize output formatter settings")
-	}
-	return s, nil
-}
-
-func (ofs *OutputFormatterSettings) CreateOutputFormatter() (formatters.OutputFormatter, error) {
-	if ofs.Output == "csv" {
-		ofs.Output = "table"
-		ofs.TableFormat = "csv"
-	} else if ofs.Output == "tsv" {
-		ofs.Output = "table"
-		ofs.TableFormat = "tsv"
-	} else if ofs.Output == "markdown" {
-		ofs.Output = "table"
-		ofs.TableFormat = "markdown"
-	} else if ofs.Output == "html" {
-		ofs.Output = "table"
-		ofs.TableFormat = "html"
-	}
-
-	var of formatters.OutputFormatter
-	if ofs.Output == "json" {
-		of = formatters.NewJSONOutputFormatter(ofs.OutputAsObjects)
-	} else if ofs.Output == "yaml" {
-		of = formatters.NewYAMLOutputFormatter()
-	} else if ofs.Output == "table" {
-		if ofs.TableFormat == "csv" {
-			csvOf := formatters.NewCSVOutputFormatter()
-			csvOf.WithHeaders = ofs.WithHeaders
-			r, _ := utf8.DecodeRuneInString(ofs.CsvSeparator)
-			csvOf.Separator = r
-			of = csvOf
-		} else if ofs.TableFormat == "tsv" {
-			tsvOf := formatters.NewTSVOutputFormatter()
-			tsvOf.WithHeaders = ofs.WithHeaders
-			of = tsvOf
-		} else {
-			of = formatters.NewTableOutputFormatter(ofs.TableFormat)
-		}
-		of.AddTableMiddleware(middlewares.NewFlattenObjectMiddleware())
-	} else if ofs.Output == "template" {
-		if ofs.TemplateFormatterSettings == nil {
-			ofs.TemplateFormatterSettings = &TemplateFormatterSettings{
-				TemplateFuncMaps: []template.FuncMap{
-					sprig.TxtFuncMap(),
-					helpers.TemplateFuncs,
-				},
-				AdditionalData: make(map[string]interface{}),
-			}
-		}
-		of = formatters.NewTemplateOutputFormatter(
-			ofs.Template,
-			ofs.TemplateFormatterSettings.TemplateFuncMaps,
-			ofs.TemplateFormatterSettings.AdditionalData,
-		)
-	} else {
-		return nil, errors.Errorf("Unknown output format: " + ofs.Output)
-	}
-
-	return of, nil
-}
-
-type TemplateSettings struct {
-	RenameSeparator string
-	UseRowTemplates bool
-	Templates       map[types.FieldName]string
-}
-
-func (tf *TemplateSettings) AddMiddlewares(of formatters.OutputFormatter) error {
-	if tf.UseRowTemplates && len(tf.Templates) > 0 {
-		middleware, err := middlewares.NewRowGoTemplateMiddleware(tf.Templates, tf.RenameSeparator)
-		if err != nil {
-			return err
-		}
-		of.AddTableMiddleware(middleware)
-	}
-
-	return nil
-}
 
 type FieldsFilterSettings struct {
 	Filters        []string
@@ -129,7 +23,6 @@ func (fff *FieldsFilterSettings) AddMiddlewares(of formatters.OutputFormatter) {
 	if len(fff.ReorderColumns) > 0 {
 		of.AddTableMiddleware(middlewares.NewReorderColumnOrderMiddleware(fff.ReorderColumns))
 	}
-
 }
 
 type SelectSettings struct {
